+++ helpers/git_helpers.sh
@@ -25,7 +25,7 @@ function checkout_main_branch {
 function cherry_pick {
   local SHAs=$*
   run_command "git cherry-pick $SHAs"
-  if [ $command_exit_status != 0 ]; then error_cherry_pick $SHAs; fi
+  if [ $? != 0 ]; then error_cherry_pick $SHAs; fi
 }
 
 # Creates a new feature branch with the given name.
@@ -70,7 +70,6 @@ function has_tracking_branch {
   fi
 }
 
-
 # Exists the application with an error message if the
 # current working directory contains uncommitted changes.
 function ensure_no_open_changes {
@@ -108,6 +107,17 @@ function error_pull_branch {
 # Returns the current branch name
 function get_current_branch_name {
   git branch | grep "*" | awk '{print $2}'
+}
+
+
+# Determines whether the given branch is ahead of main
+function is_ahead_of_main {
+  local branch_name=$1
+  if [ `git log --oneline $main_branch_name..$branch_name | wc -l` == 0 ]; then
+    echo false
+  else
+    echo true
+  fi
 }
 
 
@@ -158,7 +168,7 @@ function merge_branch {
   local branch_name=$1
   local current_branch_name=`get_current_branch_name`
   run_command "git merge $branch_name"
-  if [ $command_exit_status != 0 ]; then error_merge_branch; fi
+  if [ $? != 0 ]; then error_merge_branch; fi
 }
 
 
@@ -185,7 +195,7 @@ function pull_branch {
   if [ `has_tracking_branch` == true ]; then
     fetch_repo
     run_command "git $strategy origin/$current_branch_name"
-    if [ $command_exit_status != 0 ]; then error_pull_branch $current_branch_name; fi
+    if [ $? != 0 ]; then error_pull_branch $current_branch_name; fi
   else
     echo "Branch '$current_branch_name' has no remote branch, skipping pull of updates"
   fi
@@ -197,7 +207,7 @@ function pull_upstream_branch {
   local current_branch_name=`get_current_branch_name`
   fetch_upstream
   run_command "git rebase upstream/$current_branch_name"
-  if [ $command_exit_status != 0 ]; then error_pull_upstream_branch; fi
+  if [ $? != 0 ]; then error_pull_upstream_branch; fi
 }
 
 
@@ -242,13 +252,11 @@ function squash_merge {
   local commit_message=$2
   local current_branch_name=`get_current_branch_name`
   run_command "git merge --squash $branch_name"
-  if [ $command_exit_status != 0 ]; then error_squash_merge; fi
+  if [ $? != 0 ]; then error_squash_merge; fi
   if [ "$commit_message" == "" ]; then
-    print_command "git commit -a"
-    git commit -a
+    run_command "git commit -a"
   else
-    print_command "git commit -a -m \"$commit_message\""
-    git commit -a -m "$commit_message"
+    run_command "git commit -a -m '$commit_message'"
   fi
 }